PXC-3664: clustercheck in some cases fails to identify correct node state
https://jira.percona.com/browse/PXC-3664

This commit aligns the logic of clustercheck.sh with the logic of its
counterpart script check_pxc.sh located in percona-docker repository.
It doesn't solve the problem described in the ticket because it is
HAProxy which assumes the node is OK by default, even before the 1st
check. So the problem is related to HAProxy behavior and should be fixed
by HaProxy state file, not in PXC (there is no way to fix it on PXC or
check the script side).

Original contribution by Antonio Falzarano.
https://github.com/percona/percona-xtradb-cluster/pull/1428
Not merged directly because of GCA merge process which needs to be
applied.