PXC-3999: Cluster stuck on "Waiting for acl cache lock" with concurrent User DDL commands
https://jira.percona.com/browse/PXC-3999

Problem:
When multiple CREATE/ALTER USER statements are executed in parallel
the node can stuck with deadlock.

Cause:
This is the regression caused by commit
6cc5a28b7c4a72239b35804fc704c2ee9a32abad.
In that commit the order of acquiring acl_cache_lock and starting TOI
was changed (now acl_cache_lock is acquired first).
When CREATE/ALTER USER is executed, acl_cache_lock is acquired, then TOI
is started. So far so good, but then acl_cache_lock is released and
acquired again. This may lead to the situation when another thread
acquires acl_cache_lock and tries to start TOI, which will block as the
1st thread has active TOI. Then, if the 1st thread tries to acquire
acl_cache_lock it will end up with a deadlock.

Additionally, it has been found that CHANGE PASSWORD has wrong
acl_cache_lock and start TOI order, which could lead to the deadlock
between session changing the password and another session which executes
user alter related statement (e.g. RENAME USER)

Other paths are not affected.

Solution:
1. Extended acl_cache_lock critical session scope in mysql_create_user()
/mysql_alter_user() functions. Now the acl_cache_lock is not released in
the middle of the function keeping it in write mode.
As previously it was reacquired in read mode, this potentially affects
concurrency, but it should not be a big deal.
2. Fixed locking order in change_password() function.