PXC-4033: Assertion during execution of prepared statement
https://jira.percona.com/browse/PXC-4033

Problem:
When the prepared statement is executed in parallel to the DDL modifying
the table which prepared statement uses the server fails with assertion
saying that prepared statement transaction was aborted, so it cannot be
commited.

Cause:
When the session executing prepared statement is opening the table it
detects that table's metadata changed. Execution is aborted, internal
error is propagated up the callstack and as the transaction ended with
error, wsrep part marks it as aborted.
Error is internal, its purpose is to trigger prepared statement
re-execution, which happens, but wsrep transaction state is in aborted
state. Later when the transaction is commited, it is detected that
transaction is aborted and the assert triggers.

During the investigation I've found that for multi statement
transactions the state is also changed to aborted which causes skipping
certification (and replication) of such transaction, but local commit
happens which leads to the cluster inconsistency.

Solution:
Before re-execution of aborted prepared statement do the cleanup (call
wsrep_after_statement()).
Added assertions in wsrep-lib.