PXC-4012: Node leaves PXC cluster when running concurrent CREATE USER with password_history option
https://jira.percona.com/browse/PXC-4012

Problem:
When password history is enabled and CREATE USER is executed in
parallel, or one after another, but in the same second the replica node
leaves the cluster.

Cause:
mysql.password_history has the primary key on User, Host, and Timestamp
columns. Log_event_header carries the timestamp with one-second
resolution. The information about the microseconds part of the timestamp
is carried by Q_MICROSECONDS variable stored along with the event.
However, as Q_MICROSECONDS is needed only for a few events, there is
optimization related to THD::query_start_usec_used flag. This flag is
set only for queries that require a high-precision timestamp when the
query is processed. Then binlog event is logged and replicated (async
replication).
With Galera replication in place, we replicate DDL as TOI. It happens
at the very beginning of the query and THD::query_start_usec_used flag
is not set yet, so the microseconds part of the timestamp is not sent.

Solution:
For user's password-related queries set THD::query_start_usec_used
explicitly for the time of construction of Query_log_event which will be
replicated. This is a temporary event, only for TOI replication.