PXC-4087: BF abort for MDL conflict causes a crash when the victim thread is executing a rollback and thread pool is enabled
https://jira.percona.com/browse/PXC-4087

Problem:
The connection on node-1 having big uncommited transaction is
killed. This causes transaction on node-1 being rolled back. If during
the rollback node-2 issures another transaction which confilcts with
the one from node-1, the server asserts (debug) or crashes (release).
It only happens if thread_handling is 'pool-of-threads'.

Cause:
The proper flow is:
1. node-1 transaction (T1) is being rolled back as the consequence of
connection KILL. THD::wsrep_query_state is QUERY_EXEC, which indicates
ongoing query.
2. Transaction (T2) from node-2 is replicated. It sees that there is
conflicting T1, so T1 has to be BF-aborted.
3. T1 abort is triggered, however applier thread (T2) waits for T1's
rollback
4. Applier thread continues with T2

The problem was that in case of 'pool-of-threads' usage,
THD::wsrep_query_state was set to QUERY_IDLE instead of QUERY_EXEC
for the time of rollback. This caused logic of applier to not wait for
T1, but to delegate the rollback to the wsrep rollbacker thread.

ha_rollback_low() is executed as a consequence of connection KILL
resets/clears ha_info of rolled back SE.
In the same time rollbacker thread tries to register wsrep handlerton
in wsrep_register_hton(). It iterates over ha_infos of transaction and
tries to dereference already cleared m_ht.

Solution:
Changed how THD::wsrep_query_state is set to QUERY_EXEC/QUERY_IDLE to
include the case when the connection is killed and does rollback as the
part of error handling.