PXC-4126: Create user is forbidden when stream replication is active
https://jira.percona.com/browse/PXC-4126

Problem:
When SR is enabled, CREATE USER statement fails

Cause:
This issue is the regression caused by fix for PXC-3872 (commit 487328e781d5acd44cd94e526969874372080b51). It moved starting TOI after
open_grant_tables(). Before the fix, TOI was started at the beginning
of open_grant_tables() (so effectively, 'before' open_grant_tables()).
open_grant_tables() calls decide_logging_format(), where we make a check
if we are in local mode and doing streaming replication (SR).
wsrep_thd_is_local() is used as a part of the decision input. If we are
in TOI mode, wsrep_thd_is_local() returns 'false' (before above
mentioned commit), but if we didn't start TOI yet, it returns 'true',
so the error is returned, as local + STMT + SR is not allowed.

Solution:
As SR and TOI are independent things, instead of disabling SR for the
time TOI is executed, inform the logic in decide_logging_format() that
TOI is about to be started (THD::wsrep_TOI_preparation flag introduced)