PXC-4116: PXC stalls with parallel replication workers
Issue: MySQL parallel replication workers and galera both order commits
in a well defined order, and use locks/signals to enforce it, but they
end up generating different sequences.

Galera assigns monotonic sequence numbers in wsrep_before_prepare, at
the prepare stage of the commit, and waits for earlier commits in
wsrep_before_commit, before starting the commit.

MySQL uses the commit order in the relay log, and waits for earlier
commits later before flushing. Up to that point, workers process commits
in parallel.

If the parallel workers reach wsrep_before_prepare in a different order
than the order in the relay log, wsrep assigns sequence numbers in a
different order than MySQL, and the two logics will end up in a
deadlock:

Commit A has the higher galera seqno and waits for commit B in the
galera code, but Commit B has the higher MySQL seqno and waits for
commit A in teh server code.

Additionally, one seqno for Galera means one commit, while for MySQL it
can mean multiple commits.

Fix: we disable parallel execution of commits before calling
wsrep_before_prepare. Only the worker that is next in line to flush the
commit can start preparing the commit.

This strict exclusion is required. A more relaxed implementation, wihch
only requires workers to enter wsrep_before_prepare in the same order as
their seqnos doesn't work because of the issue described at the end:

As we can process multiple commits with the same server seqno, we can't
let other threads proceed until one worker flushed all of its commits.
If we allow that, we could assign galera seqno N+1 to the next in line
worker, and after that galera seqno N+2 to the second commit of the
current worker, resulting in the same deadlock.