PXC-4137: WSREP applier threads fail to modify read only schemas
https://jira.percona.com/browse/PXC-4137

Problem
-------
When a PXC cluster acts as the async replica to a MySQL source, and if the
cluster has a schema has marked as read only in the PXC cluster, then updates
on the schema from async source server are not handled incorrectly leading to
inconsistencies between PXC cluster nodes.

Analysis
--------
Consider the scenario where a schema has been marked as read only in the PXC
Cluster, while the schema on source server is open for writes.

Topology:

               async           PXC
        node1 -------> node2 <-----> node3

1. n1> CREATE DATABASE testdb;
2. n2> ALTER SCHEMA testdb READ ONLY = 1; -> schema will be read only on all
                                             nodes of the cluster
3. n1> CREATE TABLE testdb.t1 ..

In such a case, the CREATE TABLE query in step 3 will be applied on pxc replica
by the replication applier thread on node2 and the same will also be replicated
to node3 as TOI.

Since the replication applier threads are allowed to continue even when schema
is read only, the CREATE TABLE will be successful on node2.

However, the same operation will error out on node3 as wsrep applier threads
are not allowed to operate on read only schemas, thereby resulting in
inconsistency voting to kick-in and thereby resulting in the cluster shutdown.

NOTE:
This problem can be easily fixed by allowing the wsrep applier threads to
operate on read only schemas. However this could still result in inconsistency
if the user tries to create database objects directly on PXC nodes (node2 and
node3) as the the queries are replicated as TOI before the checks for read only
schemas are carried out.  As a result, the query can still succeed on the
replicated node while it fails on the originating node, thereby again resulting
in inconsistency followed by cluster shutdown.

node2> create table testdb.t1(i int);
ERROR 3989 (HY000): Schema 'testdb' is in read only mode.

Solution
--------
1. Allow wsrep applier threads to operate on read only schemas.
2. Add checks to disallow creation of database objects inside read only schemas
   before the query is replicated as TOI.