PXC-4163: pxc_strict_mode does not detect version mismatch
https://jira.percona.com/browse/PXC-4163

Server part of the fix.

Problem:
When in mixed-version cluster, 8.0 node becomes a donor for another 8.0
node, for the time of being donor, the node accepts writes.

Cause:
block_write_while_in_rolling_upgrade() function skips check if the node
is not in SYNCED state.

Solution:
We need to keep checking for write restriction in mixed-version cluster
for all the time, regardless of the state of the node.
The only time we have to accept writes is when the node joins
the cluster and does the initialization. Once the node is initialized
it has to reject writes even if it becomes a donor.

Galera repo pointer updated.