PXC-3047: Server hangs up if cannot connect to members PXC 8.0
https://jira.percona.com/browse/PXC-3047

Problem:
The server with a view stored in gvwstate.dat file hangs up infinitely,
when it starts, but can not connect to other nodes.

Cause:
When the node starts, it waits for primary component for
pc.wait_prim_timeout seconds. If PC is not reached, the node exits.
However, this timeout is not respected in case of the PC view restored
from the file. In such a case node just joins the cluster and waits for
PC infinitely.

Solution:
(The solution is in Galera library, server part contains only MTR test)
pc.wait_prim_timeout defaults to 30s and it may be not enough for
cluster recovery after failure. Introduce new configuration option
pc.wait_restored_prim_timeout with default of 0s.
If set to default, there is the old behavior - the node waits infinitely
if set to any other value, the node waits for requested time (sec) for
PC and exits if not reached.