PXC-3421 IMPROVES ERROR HANDLING IN WSREP_SCHEMA.CC
https://jira.percona.com/browse/PXC-3421

All the failure scenarios in wsrep_schema.cc are suffixed with a call to
finish_stmt() method at the end before erroring out. But, internally it
invokes trans_commit_stmt() which actually commits the transaction which
is unnecessary.

So, the transaction is properly rollbacked and followed by closing of
the tables.