PXC-679: Undetected state gap discovery causes server hung on shutdown
https://jira.percona.com/browse/PXC-679

Problem:
When the server detects a state gap which was not detected at galera
level it aborts itself. It is OK however, during the shutdown the server
hangs.

Cause:
This is because we call unireg_abort() from wsrep_view_handler_cb
(view_cb_), which is done from wsrep applier context. The process is
waiting for wsrep applier thread to be finished and never exits.

Solution:
1. Return WSREP_CB_FAILURE from view callback instead of calling
unireg_abort(). The failure will be detected by galera and abort will
be called.
2. Prohibit setting wsrep_start_position when wsrep provider is loaded
and server is not starting.
Backported 8.0 commit a11616faa1b8781dd6d36e3c0d25097d2c5d52a0 idea.
However, note that the original logic implemented in this commit is
wrong and is in contradiction to the warning message displayed.
wsrep_start_position variable's purpose is to tell the Joiner its start
position after SST. It is used by mysqldump SST.
The condition which prohibits setting wsrep_start_position has been
fixed.
3. Fixed wsrep_start_position default value
(SET GLOBAL wsrep_start_position=default didn't work as nullptr was
the default value instead of 00000000-0000-0000-0000-000000000000:-1
4. wsrep_start_position_basic MTR test moved to wsrep suite and fixed.
5. Added wsrep suite to the list of default suites