PXC-4258: Failed to add foreign key results in Inconsistency
https://jira.percona.com/browse/PXC-4258

Problem
-------
ALTER TABLE on a table involving a FK relationship can sometimes result
in cluster inconsistency due to different error messages.

Different messages are possible when the error message includes the name
of the intermediate temporary table that gets created during the
execution of the ALTER TABLE.

  Example:

  node-1: foreign key constraint fails (`test`.`#sql-25528d_2`, CONSTRAINT `child_ibfk_1`
  node-2: foreign key constraint fails (`test`.`#sql-22ed07_9`, CONSTRAINT `child_ibfk_1`

Solution
--------
Avoid printing the temporary table name and instead print "temp_table"
to the error message for TOI and NBO operations.