PXC-4182: Assertion 'active() == false' failed
https://jira.percona.com/browse/PXC-4182

Problem
-------
When PXC node is operating as an async replica of a MySQL server,
cleanup of wsrep context of the applier thread was not being done during
the transaction retry of the replicated transaction resulted in the
below assertion failure.

   Assertion "active() == false" in wsrep::transaction::start_transaction

Solution
--------
The applier threads now perform cleanup of wsrep context before
performing a transaction retry.