PXC-4302: GRANT statement may be replicated in a wrong way if partial_revokes=1
https://jira.percona.com/browse/PXC-4302

Problem:
If partial_revokes is enabled and the user executes on node_1 GRANT
statement without explicit AS clause, the binlog on node_2 contains
wrong event having "AS 'skip-grants user'@'skip-grants host'" which
may break async replication if node_2 is a source node for such a chain.

Cause:
If partial_revokes is enabled, AS clause is not explicitly specified
and the GRANT is on global (*.*) level, node_1 rewrites the query just
before writing the event into the binlog, adding AS clause with
the current user. This causes async replica node to explicitly see AS
clause and not doing any rewrite (replica thread works in root user
context)

However, we replicate TOI with the query being "as is" before local
execution. It triggers the rewrite path on replica node, causing event
with wrong AS clause being stored in its binlog.

Solution:
Rewrite query on source node for TOI replication in the same way as
it is rewritten before storing into the binlog.