PXC-4312: DROP EVENT IF EXISTS generates local GTID event
https://jira.percona.com/browse/PXC-4312

Problem:
When DROP EVENT IF EXISTS is executed for non existing event, the event
is binlogged with the GTID containing UUID of local server instead of
global, cluster-wide UUID.

Cause:
As per MySql documentation, DROP DATABASE IF EXISTS,
DROP TABLE IF EXISTS, and DROP VIEW IF EXISTS have to be replicated
always. Code analysis lead to the conclusion that DROP EVENT IF EXISTS
is in this group as well.
We did TOI for all besides DROP EVENT IF EXISTS. If the query is not
replicated, wsrep logic generates GTID using local server UUID instead
of cluster-wide UUID.

Solution:
Replicate DROP EVENT IF EXISTS always. This is enough for GTID to be
properly generated.