PXC-4318: long semaphore wait crash due to ha_commit_low does not commit an empty transaction
https://jira.percona.com/browse/PXC-4318

Problem:
When PXC node acts as the async replica to some master and in parallel
the row is updated on PXC node and via replication the PXC node hangs.

Cause:
If the row is updated in parallel by local session on PXC node and via
replication in the way that both updates result in the same row, and
local transaction updated the row as the first one, the replicated
transaction is not logged into the binlog.
This is because when replicated transaction commits in InooDB,
SE returns HA_ERR_RECORD_IS_THE_SAME and binlogging is skipped
(logic inside handler::ha_update_row()). So the transaction becomes
an empty transaction. In this case slave protocol directly commits
gtid through gtid_end_transaction that invokes ordered_commit causing
thread handler to register in wsrep group commit queue but since storage
engine commit is not done it would fail to unregister the said thread
handler as part of storage engine commit.

In other words, the problematic flow is:
1. Two threads in commit queue T1 and T2. T1 is a leader, T1 skips SE
2. commit_queue looks like this: T1, T2
3. wsrep_commit_queue looks like this: T1, T2
4. T1 starts processing commit_queue
5. T1 processes itself. Skips SE commit, so it is still at the front
of wsrep_commit_queue
6. T1 processes T2. It commits in SE. First it does it waits for T2
to be in front of wsrep_commit_queue -> infinite wait

Commit 730b82ac tried to solve the above situation by removing
the thread from wsrep group commit queue in wsrep_after_commit() if it
was not removed by committing to SE.
The problem of this solution is when the thread (t1) registers in
wsrep_group_commit_queue, another thread (t2) may register after it
(t1 does not have to be a leader). The queue is processed by the leader
and each thread is supposed to remove itself from
wsrep_group_commit_queue during SE commit. If t1 does not do SE commit
it remains at the front of the queue, and we block during t2 processing
in wsrep_wait_for_turn_in_group_commit().
So unregistering t1 from the queue in wsrep_after_commit() (after the
whole commit queue is done) is too late.

Solution:
If the transaction does not cause commit in SE which would cause
removal of it from wsrep_group_commit_queue, remove it explicitly.
This is a similar approach as implemented in 730b82ac but moved directly
to after processing of the thread during COMMIT stage rather than having
it after the whole stage is finished.