PXC-4336: PXC node eviction due to CHECK CONSTRAINT
https://jira.percona.com/browse/PXC-4336

Problem:
When constraint, which fails, is set on table, inconsistency voting is
kicked off and the node is evicted from the cluster.

Cause:
ALTER TABLE ... ADD CONSTRAINT ... is replicated as TOI before the
actual execution on the source node. Than it fails to execute on source
node and succeeds on replicated node. This activates inconsistency
voting protocol and the source node gets evicted from the cluster.

When the query is executed, before the actual ALTER execution, check
constraints specifications are created
(prepare_check_constraints_for_alter()).
On source node they are created, then validated, validatio fails, and
the node finishes with error.
On replicated node constraints specifications creation is skipped, so
constraints are not validated at all. It is because of the logic of
is_slave_with_master_without_check_constraints_support() function, which
forces constraints check skipping in case of source node is not
supporting such checks. The decision is based on the version of the
source node which is carried in Gtid event. Galera does not replicate
Gtid events, so this information is missing and the logic assumes that
the source node does not support constraints check.

Solution:
Allow constraints check for Galera replication.