PXC-4345: PXC 5.7.44 refresh - Q4 2023 (Fix ASAN failure)
https://jira.percona.com/browse/PXC-4345

Problem
-------
The test galera.galera_flush_table_for_export_thread_pool fails with
heap-use-after-free when ran with ASAN.

Analysis & Fix
--------------
In general, when an error occurs in the `handle_event()` funcion in
threadpool, server performs cleanup of the connection object by calling
`connection_abort()`. However the PXC specific code still tried to
access the freed memory while trying to set the query state, thus
resulting in the heap-use-after-free.

This has been fixed by changing the order of the checks. We now set the
query state first before checking if the connection needs to be
terminated.