PXC-4317: Percona XtraDB 8.0.33-25-1 Nodes 2 and 3 cannot join the cluster PXC-4237: PXC tarball install : Failed wsrep_sst_xtrabackup-v2 when adding new node
https://jira.percona.com/browse/PXC-4317
https://jira.percona.com/browse/PXC-4237

Problem
-------
On newer platforms like AlmaLinux, the readlink command used during the SST process on
joiner can fail resulting in the failure to add a new node to an
existing cluster.

Solution
--------
Added a pair of `set +e` and `set -e` respectively before and after the
readlink command so that any non-zero return value will not be
considered as a failure during the execution of the script.