PXC-4034: Using sql_log_bin=0 breaks GTID consistency
https://jira.percona.com/browse/PXC-4034

Problem:
If binlogging is enabled globally, but it is disabled in the session
scope, transactions executed within this session advance 'gtid_executed'
As transactions are not binlogged, after the node restart, 'gtid_purged'
is populated with gtids related to these transactions.
When the node acts as the source for async replica, it causes replica
to fail, because source is not able to serve not logged, but
gtid-related transactions.

Cause:
PXC, even if binlog is disabled, uses binlog infrastructure to cache
binlog events for the purpose of creation of Galera writesets. If binlog
is disabled, logging to the file is skipped.
GTID is generated during group commit binlog flush stage.
PXC flow goes through this stage if binlog is enabled generating GTIDs
for every transaction, even if later the transaction is not binlogged.

Solution:
If binlogging is disabled for the particular thread, force the execution
path as if gtid_mode was equal OFF.