PXC-4341: Executing prepared statement can abort node after FLUSH TABLES
https://perconadev.atlassian.net/browse/PXC-4341

Problem
-------
Transient errors during the execution of the DDL in a prepared statement
mode can trigger inconsistency voting resulting in node eviction from
the cluster.

Analysis & Fix
--------------
It can sometimes happen that the DDL executed as a prepared statement
may fail with transient errors like ER_NEED_REPREPARE if the table
metadata was changed between PREPARE and EXECUTE (due to FLUSH TABLES).
In such cases, the statement is re-prepared and re-executed from
Prepared_statement::execute_loop().

However, before the statement is reprepared, PXC's TOI_end() hook captures
this transient error (i.e, ER_NEED_REPREPARE) and sends it to galera thereby
triggering the inconsistenct voting and kicking this node out of the
cluster.

So, in such cases, we need to delay sending the final error to galera until
all reprepare attempts are over. This is done by attaching a temporary
Diagnostics_area to THD and resetting it later if all the required conditions
are met.

NOTE
----

The below mentioned commands reprepare the statement if there is any
metadata change (marked with CF_REEXECUTION_FRAGILE)

    SQLCOM_CREATE_TABLE
    SQLCOM_LOAD
    SQLCOM_CREATE_VIEW
    SQLCOM_UPDATE
    SQLCOM_UPDATE_MULTI
    SQLCOM_INSERT
    SQLCOM_INSERT_SELECT
    SQLCOM_DELETE
    SQLCOM_DELETE_MULTI
    SQLCOM_REPLACE
    SQLCOM_REPLACE_SELECT
    SQLCOM_SELECT
    SQLCOM_DO
    SQLCOM_CALL

Among the above commands, only SQLCOM_CREATE_TABLE and
SQLCOM_CREATE_VIEW are DDLs and thus was affected by this bug.

However in case of CREATE VIEW, table version check is performed before
the query is replicated as TOI, so CREATE TABLE was the only query that
was affected by this bug. So, a special handling has been done only for
the CREATE_TABLE query in this patch.

(cherry picked from commit ca1944fb65aacf063d32e0f067829fadfb5a3843)