PXC-4313: Statements executed in RSU mode generate local GTID events
https://perconadev.atlassian.net/browse/PXC-4313

Problem:
When the DDL is executed in RSU mode it generates binlog event and new
GTID. As this is RSU, the change is not replicated, which results with
binlog inconsistencies between PXC cluster nodes.
If PXC acts as source for async replica, the failover process doesn't
work as nodes have different binlogs.

Solution:
When DDL is executed in RSU mode:
1. Do not create binlog event
2. Do not generate GTID
3. Log a warning containing the DDL query
4. Send warning to the client