PXC-4434: Merge PS 5.7.44-50 Q2 2024 (Fix MTR failures)
https://perconadev.atlassian.net/browse/PXC-4434

Commit 83fcba60ab75c66b7e417a3c19fa63b2a43a1278 fixed the
heap-use-after-free issue by changing the order of checks performed during
the cleanup. However this resulted in a scenario where multiple threads performed
rollback of a transaction resulting in assertion failure.

This happened because THD::wsrep_query_state was set to QUERY_IDLE
instead of QUERY_EXEC at the time of rollback. This caused logic of
applier to not wait for T1, but to delegate the rollback to the wsrep
rollbacker thread.

This atch reverts the commit 83fcba60ab75c66b7e417a3c19fa63b2a43a1278
("PXC-4345: PXC 5.7.44 refresh - Q4 2023 (Fix ASAN failure)")
and fixes the original heap-use-after-free in a different way.