PXC-4499: Remove checks for prev pxb from SST script
https://perconadev.atlassian.net/browse/PXC-4499

Problem:
PXC 8.4 will be released for Ubuntu Noble. SST script allows PXC SST to
be done for the following combinations:
1. Previous LTS -> this version (8.0.x -> 8.4)
2. Previous version -> this version (8.3 -> 8.4)
4. This version -> this version (8.4 -> 8.4)

Support of path 2 would need PXB 8.3 to be released for Ubuntu Noble,
but it was not released.

On the other hand, all innovative releases are internal, so public
release needs to support only paths 1 and 3

Solution:
Remove check for prev PXB version from SST script. This way only
previous LTS and 'this version' will be a hard dependency.