PXC-4500: When innodb_thread_concurrency is set, the cluster can get stuck during SST
https://perconadev.atlassian.net/browse/PXC-4500

Problem 1:
When innodb_thread_concurrency is set and the node has a heavy user
workload, it can become stuck when it receives a CC event from Galera.

Cause:
Let's say innodb_thread_concurrency = 2.
1. We have two user threads that have been granted access to InnoDB.
They will release the InnoDB access lock after Galera certification,
but for certification, they need to acquire LocalMonitor.
2. At the same time CC happens. The Applier thread acquires LocalMonitor
and notifies the application about the new view. The application tries
to store the view in wsrep schema. Before doing that it turns off
wsrep_on thread's variable (the thread is not marked as wsrep-enabled
thread). Then, it tries to enter InnoDB. As we already have two user
threads in InnoDB and our thread is not wsrep thread, we have to wait.

The above results in a deadlock:
1. The user thread is holding InnoDB lock, waiting for LocalMonitor
2. The Applier thread is holding LocalMonitor, waiting for InnoDB lock

Solution:
The wrong condition was used to detect the wsrep applier thread in
innobase_srv_conc_enter_innodb(). The applier thread should always be
granted access. Fixed.

Problem 2:
Even after fixing Problem 1, the cluster was stuck during SST.

Cause:
The SST thread creates an SST user. For this, it needs to enter InnoDB.
We end up in a similar situation as in Problem 1. The applier thread
holds LocalMonitor and waits for SST to finish. The SST thread waits for
InnoDB. User threads hold the InnoDB lock and wait for LocalMonitor.

Solution:
Allow SST thread to enter InnoDB always.