PXC-4504: DROP PROCEDURE/FUNCTION IF EXISTS generates local GTID event
https://perconadev.atlassian.net/browse/PXC-4504

Problem:
When DROP PROCEDURE/FUNCTION IF EXISTS is executed for non existing
object, the event is binlogged with the GTID containing UUID
of local server instead of global, cluster-wide UUID.

This case is similar to the one already fixed in commit cc8ec98a
(PXC-4312)

Cause:
DROP PROCEDURE/FUNCTION IF EXISTS is not replicated if the object does
not exist, however is binlogged. If the query is not replicated,
wsrep logic generates GTID using local server UUID instead of
cluster-wide UUID.

Solution:
Replicate DROP PROCEDURE/FUNCTION IF EXISTS always.
This is enough for GTID to be properly generated.