PXC-4498: PXC node "stalls" and causes the cluster to stop responding
https://perconadev.atlassian.net/browse/PXC-4498
https://perconadev.atlassian.net/browse/PXC-4390

Removed wsrep_group_commit_queue.

The queue works only for variables used for debug assertions. The actual
write to sys_header is done without the protection of the queue.

As the queue is not enforcing proper order of xid storing in sys_header,
but was in the past and still is the source of deadlocks - removing.