PXC-4585: Clone SST on joiner waits indefinitely if the donor node is killed while SST.
https://perconadev.atlassian.net/browse/PXC-4585

Problem 1:
If the donor node is gone after SST process is started, the joiner waits
indefinitely.

Problem 2:
If the joiner is killed during the SST process, the clone instance
remains alive in background.

Solution:
Introduced sst progress monitor on Joiner side. After inititiating
clone transfer it monitors a few things:
1. If the clone instance is still alive (if clone transfer is finished)
2. If the parent process is still alive. Actually this mechanism is
redundant, because if partent process dies, it should terminate its
children.
3. If there is any data transfer progress. If it stalls for
sst-idle-timeout (default 120 sec), SST process is terminated.

If we detect that SST has to be terminated (timeout, parent process
died) the script exits and the cleanup is done in EXIT handler.

If the parent process is terminated, it should terminate its children
as well. It termination is handle in TERM handler.

The similar handling is for netcat process.

Additionally converted all double brackets [[ ... ]] into single
brackets for consistency.